requirements.txt
openai>=1.12.0
pydantic>=2.6.0
python-dotenv
pandas
loguru
fpdf2
requests